+++ jolt-core/src/jolt/vm/mod.rs
@@ -539,4 +539,3 @@ pub mod ram_read_write_checking;
 pub mod registers;
 pub mod registers_read_write_checking;
 pub mod rv32i_vm;
-pub mod state_manager;
